More precise shop locations
Changed the method in FindShops so that it could return more precise locations from the json response. However it has to iterate through each json object from each seperate request (Using this for now, but will try to find a way to add the responses to a single object later on).