Changed the 'SELECT' methods to match the current DB
Added new methods to be able to set queries with the new DB tables. Haven't removed the olds methods since we haven't inserted new data to the new tables. Once we have that we can test and see if the results we get are the same as the ones we've gotten from the old methods