Merge Saras-new into databse-change-new (#65)
* More precise shop locations

Changed the method in FindShops so that it could return more precise locations from the json response. However it has to iterate through each json object from each seperate request (Using this for now, but will try to find a way to add the responses to a single object later on).

* Changed the 'SELECT' methods to match the current DB

Added new methods to be able to set queries with the new DB tables. Haven't removed the olds methods since we haven't inserted new data to the new tables. Once we have that we can test and see if the results we get are the same as the ones we've gotten from the old methods

* fixed type

* Removed old methods

* Changed dbcontroller methods to purchase type

Also removed the methods that require the type of the product as the argument

* Changed other methods to type of Purchase